bene

+++ server/inferences/detect-defects.py
@@ -5,17 +5,11 @@ import numpy as np
 import subprocess
 
 def load_model(model_path):
-    """
-    Load the ONNX model.
-    """
     if not os.path.isfile(model_path):
         raise ValueError(f"Model file does not exist: {model_path}")
     return onnxruntime.InferenceSession(model_path)
 
 def preprocess_image(image, input_shape):
-    """
-    Preprocess the image for inference.
-    """
     if len(input_shape) != 4 or input_shape[1] != 3:
         raise ValueError("input_shape is not valid.")
     
@@ -27,58 +21,19 @@ def preprocess_image(image, input_shape):
     return batched_image, resized_image
 
 def infer(session, input_data):
-    """
-    Perform inference on the input data.
-    """
     input_name = session.get_inputs()[0].name
     outputs = session.run(None, {input_name: input_data})
     return outputs
 
-
 def clear_output_folder(folder_path):
     """
-    Clear the contents of the output folder.
+    Clear the contents of the output folder, removing only files that do not contain the word 'vin' in their name.
     """
     for filename in os.listdir(folder_path):
         file_path = os.path.join(folder_path, filename)
-        if os.path.isfile(file_path):
+        if os.path.isfile(file_path) and 'vin' not in filename.lower():
             os.remove(file_path)
-
-
-def postprocess(outputs, original_image, confidence_threshold):
-    """
-    Postprocess the inference results to determine if there are defects
-    and draw bounding boxes.
-    """
-    dets, labels, masks = outputs
-    num_dets = dets.shape[1]
-    has_defect = False
-    highest_confidence = 0
-
-    # Create a copy of the original image to draw bounding boxes
-    image_with_boxes = original_image.copy()
-
-    for i in range(num_dets):
-        det = dets[0, i]  # Assuming dets[0, i] contains detection information
-        confidence = det[4]  # Assuming confidence level is at the fifth place (index 4)
-
-        if confidence > confidence_threshold:
-            has_defect = True
-            highest_confidence = max(highest_confidence, confidence)
-
-            x1, y1, x2, y2 = int(det[0]), int(det[1]), int(det[2]), int(det[3])
-            width = x2 - x1 
-
-            # Draw the bounding box
-            cv2.rectangle(image_with_boxes, (x1, y1), (x2, y2), (0, 255, 0), 2)
-            
-            # Add the text below the bounding box
-            text = f"Conf: {confidence:.2f} Width:{width}px"
-            text_size, _ = cv2.getTextSize(text, cv2.FONT_HERSHEY_SIMPLEX, 0.5, 2)
-            text_w, text_h = text_size
-            cv2.putText(image_with_boxes, text, (x1, y2 + text_h + 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
-
-    return has_defect, highest_confidence, image_with_boxes
+            print(f"Deleted {filename} from {folder_path}")
 
 def clear_input_folder(folder_path):
     """
@@ -89,12 +44,36 @@ def clear_input_folder(folder_path):
         if os.path.isfile(file_path):
             os.remove(file_path)
 
+def postprocess(outputs, original_image, confidence_threshold):
+    dets, labels, masks = outputs
+    num_dets = dets.shape[1]
+    has_defect = False
+    highest_confidence = 0
+
+    image_with_boxes = original_image.copy()
+
+    for i in range(num_dets):
+        det = dets[0, i]
+        confidence = det[4]
+
+        if confidence > confidence_threshold:
+            has_defect = True
+            highest_confidence = max(highest_confidence, confidence)
+
+            x1, y1, x2, y2 = int(det[0]), int(det[1]), int(det[2]), int(det[3])
+            width = x2 - x1 
+
+            cv2.rectangle(image_with_boxes, (x1, y1), (x2, y2), (0, 255, 0), 2)
+            text = f"Conf: {confidence:.2f} Width:{width}px"
+            text_size, _ = cv2.getTextSize(text, cv2.FONT_HERSHEY_SIMPLEX, 0.5, 2)
+            text_w, text_h = text_size
+            cv2.putText(image_with_boxes, text, (x1, y2 + text_h + 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
+
+    return has_defect, highest_confidence, image_with_boxes
+
 def analyze_images_in_folder(folder_path, model_path, input_shape, output_folder, confidence_threshold):
-    """
-    Analyze all images in a folder and save images with bounding boxes in a new folder.
-    """
     session = load_model(model_path)
-    defects_found = False  # Flag to track if any defects were found
+    defects_found = False
 
     for filename in os.listdir(folder_path):
         if 'vin' not in filename.lower() and filename.lower().endswith(('.png', '.jpg', '.jpeg', '.bmp', '.tiff')):
@@ -109,7 +88,7 @@ def analyze_images_in_folder(folder_path, model_path, input_shape, output_folder
             has_defect, highest_confidence, image_with_boxes = postprocess(outputs, original_image, confidence_threshold)
 
             if has_defect:
-                defects_found = True  # Set flag to True if a defect is found
+                defects_found = True
                 base, ext = os.path.splitext(filename)
                 confidence_str = f"{highest_confidence:.2f}"
                 output_filename = f"{base}{ext}"
@@ -117,10 +96,15 @@ def analyze_images_in_folder(folder_path, model_path, input_shape, output_folder
                 cv2.imwrite(output_path, image_with_boxes)
                 print(f"Saved {output_filename} to {output_folder}")
 
-            os.remove(image_path)
-            print(f"Deleted {filename} from {folder_path}")
+            try:
+                os.remove(image_path)
+                print(f"Deleted {filename} from {folder_path}")
+            except FileNotFoundError:
+                print(f"File not found: {image_path}")
+            except Exception as e:
+                print(f"Error deleting file {image_path}: {e}")
 
-    return defects_found  # Return the flag indicating whether defects were found
+    return defects_found
 
 def main():
     input_dir = "defects/scratches"
@@ -129,17 +113,14 @@ def main():
     output_folder = './output' 
     confidence_threshold = 0.85
 
-    # Clear the output folder before starting the analysis
     clear_output_folder(output_folder)
 
     defects_found = analyze_images_in_folder(input_dir, model_path, input_shape, output_folder, confidence_threshold)
 
     if defects_found:
-        # Execute the shell command if defects were found
         command = "cd /Users/gelso/workspace/PoC/server/inferences && python3 detect-vin.py"
         subprocess.run(command, shell=True)
     else:
-        # Clear the input directory if no defects were found
         clear_input_folder(input_dir)
         print(f"No defects found. Cleared input directory: {input_dir}")
